backend/requirements.txt
Django>=4.2
djangorestframework
dj-database-url
python-dotenv
psycopg2-binary
django-cors-headers
django-filter
Pillow
